Update code styling and packages version
This commit contains changes related to code formatting and updated package versions. With the updates in .prettierrc and .eslintrc.js, the tabs are changed from 2 spaces to 4 for better code readability.

The dependencies in package.json and package-lock.json files are updated. This is to ensure the application is running with the latest package versions for better performance and security.

Also, minor refactoring in main.ts file to replace hardcoded port number with a constant. This would allow for easier updates if the port number needs to be changed in the future.